pyproject.toml
[tool.poetry]
name = "ToDoList"
version = "0.1.0"
description = ""
authors = ["Leonardo Graciano <lgraciano@raioss.net>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.8"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
